+++ noodles-csi/src/io.rs
@@ -6,10 +6,11 @@ mod indexed_record;
 mod indexed_records;
 mod query;
 pub mod reader;
+pub(crate) mod writer;
 
 pub use self::{
     filter_by_region::FilterByRegion, indexed_reader::IndexedReader, indexed_record::IndexedRecord,
-    indexed_records::IndexedRecords, query::Query, reader::Reader,
+    indexed_records::IndexedRecords, query::Query, reader::Reader, writer::Writer,
 };
 
 pub(crate) static MAGIC_NUMBER: &[u8] = b"CSI\x01";